backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# 이미지 처리
pillow>=10.0.0
rembg>=2.0.62
numpy>=1.24.3
onnxruntime>=1.15.0

# Stable Diffusion
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
# xformers>=0.0.20  # 메모리 최적화 (선택사항) - Windows Long Path 문제로 주석 처리

# ControlNet
controlnet-aux>=0.0.1

# Gemini API (REST 직접 호출)
aiohttp>=3.9.0

# 기타
python-dotenv==1.0.0
requests==2.31.0
